@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: "3.5.4"
 
 dependencies:
   flutter:
@@ -27,15 +27,17 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.3
-  intl: ^0.16.1
-  sqflite: ^1.3.1+1
-  auto_size_text: ^2.1.0
-  path: ^1.7.0
-  flutter_local_notifications: ^3.0.1+2
-  timezone: ^0.5.9
-  animated_text_kit: ^2.5.3
-  animated_widgets: ^1.0.6
+  cupertino_icons: ^1.0.8
+  intl: ^0.19.0
+  sqflite: ^2.3.3+1
+  auto_size_text: ^3.0.0
+  path: ^1.9.0
+  flutter_local_notifications: ^18.0.1
+  timezone: ^0.9.3
+  animated_text_kit: ^4.2.2
+  animated_widgets: ^1.1.0
+  permission_handler: ^11.3.1
+
 dev_dependencies:
   flutter_test:
     sdk: flutter
