@@ -1,14 +1,15 @@
 import 'dart:io';
 
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/cupertino.dart';
 
 class PlatformFlatButton extends StatelessWidget {
-  final Function handler;
+  final VoidCallback? handler;
   final Widget buttonChild;
   final Color color;
 
-  PlatformFlatButton({@required this.buttonChild,@required this.color,@required this.handler});
+  PlatformFlatButton(
+      {required this.buttonChild, required this.color, required this.handler});
 
   @override
   Widget build(BuildContext context) {
@@ -19,12 +20,15 @@ class PlatformFlatButton extends StatelessWidget {
             onPressed: this.handler,
             borderRadius: BorderRadius.circular(15.0),
           )
-        : FlatButton(
-            color: this.color,
+        : TextButton(
+            style: TextButton.styleFrom(
+              backgroundColor: this.color,
+              shape: RoundedRectangleBorder(
+                borderRadius: BorderRadius.circular(15.0),
+              ),
+            ),
             child: this.buttonChild,
             onPressed: this.handler,
-            shape: RoundedRectangleBorder(
-                borderRadius: BorderRadius.circular(15.0)),
           );
   }
 }
