@@ -8,12 +8,12 @@ class PlatformSlider extends StatelessWidget {
   final Color color;
 
   PlatformSlider(
-      {@required this.value,
-      @required this.handler,
-      @required this.color,
-      @required this.max,
-      @required this.min,
-      @required this.divisions});
+      {required this.value,
+      required this.handler,
+      required this.color,
+      required this.max,
+      required this.min,
+      required this.divisions});
 
   @override
   Widget build(BuildContext context) {
