@@ -1,17 +1,31 @@
-import 'package:flutter/cupertino.dart';
+import 'dart:io';
+
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:medicine/screens/add_new_medicine/add_new_medicine.dart';
 import 'package:medicine/screens/home/home.dart';
+import 'package:permission_handler/permission_handler.dart';
 import './screens/welcome/welcome.dart';
 
-void main() {
+Future<void> requestNotificationPermission() async {
+  if (Platform.isAndroid) {
+    PermissionStatus permission = await Permission.notification.request();
+    if (permission.isDenied || permission.isRestricted) {
+      print('Notification permission denied');
+    }
+  }
+}
+
+void main() async {
+  WidgetsFlutterBinding.ensureInitialized();
+
+  await requestNotificationPermission();
+
   runApp(MedicineApp());
   SystemChrome.setSystemUIOverlayStyle(SystemUiOverlayStyle(
-    systemNavigationBarColor: Colors.black.withOpacity(0.05),
-    statusBarColor: Colors.black.withOpacity(0.05),
-    statusBarIconBrightness: Brightness.dark
-  ));
+      systemNavigationBarColor: Colors.black.withOpacity(0.05),
+      statusBarColor: Colors.black.withOpacity(0.05),
+      statusBarIconBrightness: Brightness.dark));
 }
 
 class MedicineApp extends StatelessWidget {
@@ -22,17 +36,17 @@ class MedicineApp extends StatelessWidget {
           fontFamily: "Popins",
           primaryColor: Color.fromRGBO(7, 190, 200, 1),
           textTheme: TextTheme(
-              headline1: ThemeData.light().textTheme.headline1.copyWith(
+              displayLarge: ThemeData.light().textTheme.displayLarge?.copyWith(
                     fontWeight: FontWeight.w700,
                     fontSize: 38.0,
                     fontFamily: "Popins",
                   ),
-              headline5: ThemeData.light().textTheme.headline1.copyWith(
+              headlineSmall: ThemeData.light().textTheme.displayLarge?.copyWith(
                     fontWeight: FontWeight.w400,
                     fontSize: 17.0,
                     fontFamily: "Popins",
                   ),
-              headline3: ThemeData.light().textTheme.headline3.copyWith(
+              displaySmall: ThemeData.light().textTheme.displaySmall?.copyWith(
                     fontWeight: FontWeight.w600,
                     fontSize: 20.0,
                     fontFamily: "Popins",
