@@ -1,6 +1,5 @@
 import 'dart:math';
 
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_local_notifications/flutter_local_notifications.dart';
 import 'package:intl/intl.dart';
@@ -21,32 +20,27 @@ class AddNewMedicine extends StatefulWidget {
 }
 
 class _AddNewMedicineState extends State<AddNewMedicine> {
-  FlutterLocalNotificationsPlugin flutterLocalNotificationsPlugin;
+  late FlutterLocalNotificationsPlugin flutterLocalNotificationsPlugin;
 
   final _scaffoldKey = GlobalKey<ScaffoldState>();
   final Snackbar snackbar = Snackbar();
 
-  //medicine types
+  //medicines types
   final List<String> weightValues = ["pills", "ml", "mg"];
 
   //list of medicines forms objects
   final List<MedicineType> medicineTypes = [
     MedicineType("Syrup", Image.asset("assets/images/syrup.png"), true),
-    MedicineType(
-        "Pill", Image.asset("assets/images/pills.png"), false),
-    MedicineType(
-        "Capsule", Image.asset("assets/images/capsule.png"), false),
-    MedicineType(
-        "Cream", Image.asset("assets/images/cream.png"), false),
-    MedicineType(
-        "Drops", Image.asset("assets/images/drops.png"), false),
-    MedicineType(
-        "Syringe", Image.asset("assets/images/syringe.png"), false),
+    MedicineType("Pill", Image.asset("assets/images/pills.png"), false),
+    MedicineType("Capsule", Image.asset("assets/images/capsule.png"), false),
+    MedicineType("Cream", Image.asset("assets/images/cream.png"), false),
+    MedicineType("Drops", Image.asset("assets/images/drops.png"), false),
+    MedicineType("Syringe", Image.asset("assets/images/syringe.png"), false),
   ];
 
   //-------------Pill object------------------
   int howManyWeeks = 1;
-  String selectWeight;
+  late String selectWeight;
   DateTime setDate = DateTime.now();
   final TextEditingController nameController = TextEditingController();
   final TextEditingController amountController = TextEditingController();
@@ -107,8 +101,8 @@ class _AddNewMedicineState extends State<AddNewMedicine> {
                   "Add Pills",
                   style: Theme.of(context)
                       .textTheme
-                      .headline3
-                      .copyWith(color: Colors.black),
+                      .displaySmall
+                      ?.copyWith(color: Colors.black),
                 )),
               ),
               SizedBox(
@@ -305,17 +299,20 @@ class _AddNewMedicineState extends State<AddNewMedicine> {
 
   //--------------------------------------SAVE PILL IN DATABASE---------------------------------------
   Future savePill() async {
-    //check if medicine time is lower than actual time
+    //check if medicines time is lower than actual time
     if (setDate.millisecondsSinceEpoch <=
         DateTime.now().millisecondsSinceEpoch) {
       snackbar.showSnack(
-          "Check your medicine time and date", _scaffoldKey, null);
+          context, "Check your medicines time and date", _scaffoldKey, null);
     } else {
       //create pill object
       Pill pill = Pill(
+          id: Random().nextInt(10000000),
           amount: amountController.text,
           howManyWeeks: howManyWeeks,
-          medicineForm: medicineTypes[medicineTypes.indexWhere((element) => element.isChoose == true)].name,
+          medicineForm: medicineTypes[medicineTypes
+                  .indexWhere((element) => element.isChoose == true)]
+              .name,
           name: nameController.text,
           time: setDate.millisecondsSinceEpoch,
           type: selectWeight,
@@ -326,13 +323,16 @@ class _AddNewMedicineState extends State<AddNewMedicine> {
         dynamic result =
             await _repository.insertData("Pills", pill.pillToMap());
         if (result == null) {
-          snackbar.showSnack("Something went wrong", _scaffoldKey, null);
+          snackbar.showSnack(context, "Something went wrong", _scaffoldKey, null);
           return;
         } else {
           //set the notification schneudele
           tz.initializeTimeZones();
           tz.setLocalLocation(tz.getLocation('Europe/Warsaw'));
-          await _notifications.showNotification(pill.name, pill.amount + " " + pill.medicineForm + " " + pill.type, time,
+          await _notifications.showNotification(
+              pill.name,
+              pill.amount + " " + pill.medicineForm + " " + pill.type,
+              time,
               pill.notifyId,
               flutterLocalNotificationsPlugin);
           setDate = setDate.add(Duration(milliseconds: 604800000));
@@ -341,7 +341,7 @@ class _AddNewMedicineState extends State<AddNewMedicine> {
         }
       }
       //---------------------------------------------------------------------------------------
-      snackbar.showSnack("Saved", _scaffoldKey, null);
+      snackbar.showSnack(context, "Saved", _scaffoldKey, null);
       Navigator.pop(context);
     }
   }
