@@ -1,17 +1,25 @@
-
-
 import 'package:path/path.dart';
 import 'package:sqflite/sqflite.dart';
 
-class PillsDatabase{
+class PillsDatabase {
 
-  setDatabase() async{
+  Future<Database> setDatabase() async {
     String databasePath = await getDatabasesPath();
-    String path = join(databasePath,"pills_db");
-    Database database = await openDatabase(path,version: 1,onCreate: (Database db,int version)async{
-      await db.execute("CREATE TABLE Pills (id INTEGER PRIMARY KEY, name TEXT, amount TEXT, type TEXT, howManyWeeks INTEGER, medicineForm TEXT, time INTEGER, notifyId INTEGER)");
+    String path = join(databasePath, "pills_db");
+    Database database = await openDatabase(path, version: 1, onCreate: (Database db, int version) async {
+      await db.execute('''
+        CREATE TABLE Pills (
+          id INTEGER PRIMARY KEY AUTOINCREMENT, 
+          name TEXT, 
+          amount TEXT, 
+          type TEXT, 
+          howManyWeeks INTEGER, 
+          medicineForm TEXT, 
+          time INTEGER, 
+          notifyId INTEGER
+        )
+      ''');
     });
     return database;
   }
-
 }
