@@ -1,16 +1,18 @@
 import 'package:flutter/material.dart';
 
 class Snackbar {
-  void showSnack(String message, GlobalKey<ScaffoldState> _scaffoldKey,
-          Function undo) =>
-      _scaffoldKey.currentState.showSnackBar(
-        SnackBar(
-          content: Text(message),
-          action: undo != null ? SnackBarAction(
-            textColor: Theme.of(_scaffoldKey.currentState.context).primaryColor,
-            label: "Undo",
-            onPressed: () => undo,
-          ):null,
-        ),
-      );
+  void showSnack(BuildContext context, String message, GlobalKey<ScaffoldState> scaffoldKey, VoidCallback? undo) {
+    ScaffoldMessenger.of(context).showSnackBar(
+      SnackBar(
+        content: Text(message),
+        action: undo != null
+            ? SnackBarAction(
+          textColor: Theme.of(context).primaryColor,
+          label: "Undo",
+          onPressed: undo,
+        )
+            : null,
+      ),
+    );
+  }
 }
